# Generate generics instead of decoders with macros