--- a/src/test/thrift/test.thrift
+++ b/src/test/thrift/test.thrift
@@ -85,5 +85,5 @@ struct ListItem {
 }
 
 struct StructWithOptional {
-  1: optional string s;
+  1: optional string optionalField;
 }