This PR replaces the macros which generated circe Encoder and Decoders for scrooge-generated classes for thrift structs and unions with new macros which instead generate shapeless LabelledGeneric values. These values match the values that shapeless would generate for the equivalent case classes (mostly; there are some changes to match the existing behaviour with default values), which allows us to use circe’s semi-auto derivation to get the encoders and decoders.

This has the advantage that we no longer need to maintain as much macro complexity, and we can now create encoders/decoders for recursive struct hierarchies (either direct or indirect).

Hopefully this will also mean we can migrate to scala 3 more easily, as our macros are less complicated and more connected to shapeless, which we hear scala 3 macros are inspired by.

Unfortunately, circe’s fully automatic derivation no longer works with this approach, which means this is a breaking change for consumers of the library.

## Implementation notes

- To get it working I had to use whitebox macros instead of blackbox macros, but unfortunately the existing macros don’t work when they’re whitebox, so I’ve split the new macros into their own file (I can’t see how to have whitebox and blackbox macros in the same file).

  - Whitebox implicit def macros [allow refining of the result type by the macro](https://docs.scala-lang.org/overviews/macros/blackbox-whitebox.html). I.e., the implicit search can begin for a `LabelledGeneric.Aux[Union1, R]` where `R` is completely unconstrained, and once the macro has run the result can be examined to refine this type to `LabelledGeneric.Aux[Union1, Union1Repr]`, where `Union1Repr` is a specific HList type (which implicit search can then find a circe macro for to generate a decoder). 

- To match the existing decoder behaviour, any struct field with a default parameter is allowed to be missing from the json, in which case the default value will be used.
  
- Fezziwig now serialises missing optional values as explicit nulls, rather than missing fields. This changes the contract with consumers of the library, but consumers can use deepDropNullValues to exclude nulls, which matches how circe is used elsewhere.

- I suspect there’s a way to replace these macros altogether with shapeless-style implicits, but I don’t know how yet and this works.


## How to test

This change is pretty thoroughly tested by the repository tests, as we’ve added some while it was worked on.

## How can we measure success?

- Recursive structs can now be used with the library
- Hopefully, migration to scala 3 in the future will be easier

## Have we considered potential risks?

### Semi-auto derivation

I believe the main risk is that semi-auto derivation will be more of a pain for consumers of the library than we anticipate.

I’ve made the necessary changes to content-api-models to work with this PR in [content-api-models PR 239](https://github.com/guardian/content-api-models/pull/239), and I don’t think it’s _too_ bad.

### Compile times

It’s hard to analyse the effect these changes will have on compile times: the generated macros are much smaller for nested structs, but more implicit searching is required, and I think whitebox macros are probably less performant than blackbox macros?

I’ll try to make some measurements for content-api-models!

### Non-isomorphism

I think treating default values the way we do also means that the generated LabelledGeneric is not quite an isomorphism: if you start with a thrift struct with an optional field which has a default value, and convert an instance of that struct where the field’s value is unspecified to the generic representation and back, you will now have a thrift struct where the value is specified and matches the default value.

This may cause undesirable or confusing behaviour if using the generic instance outside of json, for example if using shapeless’ lenses or zippers.