@emdash-ie 2024-02-19
Note that this PR targets the branch for pull 49, as I added some tests there and then copied them onto this branch.