@davidfurey 2024-02-20
Are these two points in the description the same?

> This method no longer suppresses nulls in the output, but I believe that’s possible for consumers using a circe option.

> Fezziwig now serialises missing optional values as explicit nulls, rather than missing fields. This changes the contract with consumers of the library, but consumers can use deepDropNullValues to exclude nulls, which matches how circe is used elsewhere.