@emdash-ie 2024-02-20
The latest commit removes the necessity for semi-auto derivation, which should make this less painful to use.