@emdash-ie 2024-03-04
I had a look at the effect of these changes on compilation times, by comparing compile times on [content-api-models](https://github.com/guardian/content-api-models) before ([commit d77f319f](https://github.com/guardian/content-api-models/commit/d77f319f)) and after ([commit d5bbf4646](https://github.com/guardian/content-api-models/commit/d5bbf4646568f8f109d07e3bfb78ff3ca03342d6)).

I made a small number of measurements and found a decrease of ~10% in compilation times. More measurements would give more confidence this can’t be random variation, but I did vary the order I took the measurements in to try and make sure it wasn’t down to CPU warmup or etc.

## `sbt compile`

I ran `sbt clean > /dev/null; time sbt compile > /dev/null`, and got the following outputs:

- before: 
  - 123.31s
  - 121.31s
- after: 
  - 108.10s
  - 108.83s

So we have a speedup of 11% in compilation! (At the cost of having to semi-automatically derive the encoders and decoders)

## `sbt +compile`

I also tried including cross-compilation to see if that would make a difference: for these measurements I ran `sbt clean > /dev/null; time sbt +compile > /dev/null`.

- before:
  - 219.46s
- after:
  - 196.98s

This shows a speedup of 10%.

## `sbt +test`

For these measurements I actually ran the tests: `sbt clean; time sbt "project content-api-models-json; +test" > /dev/null`.

- before:
  - 242.57s
  - 231.14s
- after:
  - 204.71s
  - 195.15s

Speedup here is 15%.